\name{ALassoSurvIC-internal}
\alias{ALassoSurvIC-internal}
\alias{canon}
\alias{fun_arglist}
\alias{fun_bic}
\alias{fun_cov_parallel}
\alias{fun_covij}
\alias{fun_dq}
\alias{fun_ddq}
\alias{fun_est_parallel}
\alias{fun_ew}
\alias{fun_less}
\alias{fun_penSurvIC}
\alias{fun_unpenSurvIC}
\alias{fun_shooting_algorithm}
\alias{fun_updatelambda}
\alias{log_penlikelihood}
\alias{fun_hcross}
\alias{fun_subless}
\alias{fun_sublr}
\title{Internal functions for the ALassoSurvIC package}
\description{
Internal functions for the ALassoSurvIC package.
}
